Adicionando função para cadastrar produtos e pessoas

+++ LocadoraLocal/LocadoraLocal/src/main/java/com/locadoralocal/LocadoraLocal/domain/produtos/Filmes.java
@@ -1,13 +1,13 @@
 package com.locadoralocal.LocadoraLocal.domain.produtos;
 
 public class Filmes extends Produto {
-	private double nota;
+	private float nota;
 
-	public double getNota() {
+	public float getNota() {
 		return nota;
 	}
 
-	public void setNota(double nota) {
+	public void setNota(float nota) {
 		this.nota = nota;
 	}
 }
